fix(build): match LXW_WARN_FORMAT1 format specifier to zend_long argument
Reported as #556 on Alpine 3.22 + PHP 8.1.33 (gcc 14): three call sites
in kernel/excel.c (setColumn, setRow, defaultRowOptions) pass a
zend_long (`long int`) value to `LXW_WARN_FORMAT1("...'%d'...")`. Newer
gcc treats this -Wformat= warning as an error under Alpine's hardening
defaults, breaking `pecl install`.

Switch the format specifier to %ld and cast the argument to long so the
type matches on every supported platform without depending on
sizeof(int) == sizeof(long).

+++ kernel/excel.c
@@ -1242,7 +1242,7 @@ PHP_METHOD(vtiful_xls, setColumn)
     ZVAL_COPY(return_value, getThis());
 
     if (level < 0 || level > 7) {
-        LXW_WARN_FORMAT1("outline level must be in 0..7 range, '%d' given.", level);
+        LXW_WARN_FORMAT1("outline level must be in 0..7 range, '%ld' given.", (long)level);
         level = 0;
     }
 
@@ -1287,7 +1287,7 @@ PHP_METHOD(vtiful_xls, setRow)
     ZVAL_COPY(return_value, getThis());
 
     if (level < 0 || level > 7) {
-        LXW_WARN_FORMAT1("outline level must be in 0..7 range, '%d' given.", level);
+        LXW_WARN_FORMAT1("outline level must be in 0..7 range, '%ld' given.", (long)level);
         level = 0;
     }
 
@@ -1384,7 +1384,7 @@ PHP_METHOD(vtiful_xls, defaultRowOptions)
     ZVAL_COPY(return_value, getThis());
 
     if (level < 0 || level > 7) {
-        LXW_WARN_FORMAT1("outline level must be in 0..7 range, '%d' given.", level);
+        LXW_WARN_FORMAT1("outline level must be in 0..7 range, '%ld' given.", (long)level);
         level = 0;
     }
